requirements.txt
google-cloud-aiplatform>=1.36.0
google-cloud-ai-generativelanguage>=0.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0 
